refactor : change page location
